debug

+++ main.py
@@ -105,7 +105,7 @@ def create_app(config=None):
 
    @app.route("/zone/<ACZone>/targetTemperature")
    def ACZoneSetTemp(ACZone):
-      ACValue = request.args.get('value', default = 24, type = int)
+      ACValue = request.args.get('value', default = 24, type = float)
       urlString = myPlaceUrl + '/setAircon?json={"ac1":{"zones":{"z' + ACZone + '":{"setTemp":' + ACValue + '}}}}'
       requests.get(url = urlString)
       return "ok"
